requirements.txt
torch
torchvision
pyyaml
tensorboard
pillow
lmdb
tqdm
numpy
